\name{forestFloor-package}
\alias{forestFloorPackage}
\docType{package}
\title{
forestFloor: visualize the random forest model structure
}
\description{
forestFloor visualizes randomForests models(RF). Package enables users to understand a non-linear, regression problem or a binary classification problem through RF. Any model can be separated into a series of main effect and interactions with the concept of feature contributions.
}
\details{
\tabular{ll}{
Package: \tab forestFloor\cr
Type: \tab Package\cr
Version: \tab 1.9\cr
Date: \tab 2015-12-25\cr
License: \tab GPL-2\cr
}
}

\author{
Soren Havelund Welling
}
\references{
Interpretation of QSAR Models Based on Random Forest Methods, http://dx.doi.org/10.1002/minf.201000173 \cr
Interpreting random forest classification models using a feature contribution method, http://arxiv.org/abs/1312.1121 \cr
}

\keyword{ multivariate }
\keyword{ models }
\keyword{ non-linear }
\keyword{ robust}